+++ inst/tinytest/test_mine_text.R
@@ -3,23 +3,11 @@ library("tinytest")
 library("checkmate")
 using("checkmate")
 
-# test that all publications are accounted for
-m <- mine_text(pubs, components = c("title", "abstract", "annotation", "citation"))
-is <- rownames(pubs) %in% colnames(m)
-all(is) |> expect_true()
-
-# test that when all the text components are NA an empty table is returned
-is <- is.na(pubs$annotation)
-m <- mine_text(pubs[is, ], components = "annotation")
-expect_true(nrow(m) == 0L)
-
 # test that 3-grams result in 3 word tokens, requires JAVA system dependency
-is <- Sys.getenv("JAVA_HOME") |> test_directory_exists(access = "r")
-if (is) {
-  m <- mine_text(pubs, components = "title", ngmin = 3L)
-  nwords <- rownames(m) |>
+if (test_directory_exists(Sys.getenv("JAVA_HOME"), access = "r")) {
+  d <- mine_text(docs = pubs$title, ngmin = 3L)
+  nwords <- as.character(d$term) |>
     strsplit(" ") |>
     vapply(FUN = length, FUN.VALUE = integer(1))
-  all(nwords == 3L) |>
-    expect_true()
+  all(nwords == 3L) |> expect_true()
 }